t103

+++ 0x02-functions_nested_loops/103-fibonacci.c
@@ -21,7 +21,7 @@ int main(void)
 		i++;
 		m = m + n;
 	}
-	printf("%ld\n", s);
+	printf("%ld\n", (unsigned long)s);
 
 	return (0);
 }
